% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_contrast.R
\name{aes_contrast}
\alias{aes_contrast}
\title{An auto-contrast colour aesthetic}
\usage{
aes_contrast(..., dark = "#121B24FF", light = "#FFFFFFFF")
}
\arguments{
\item{...}{Provided to force user argument naming etc.}

\item{dark}{A dark colour.}

\item{light}{A light colour.}
}
\value{
An aesthetic
}
\description{
A colour aesthetic for annotation that automatically contrasts with fill. Can be spliced into \link[ggplot2:aes]{ggplot2::aes} with \link[rlang:splice-operator]{rlang::!!!}.
}
\examples{
library(ggplot2)

data.frame(x = c("A", "B", "C"), y = c(3, 4, 5)) |>
  ggplot(aes(x = x, y = y, fill = x)) +
  geom_col(position = position_dodge2(preserve = "single"), width = 0.75, ) +
  scale_fill_manual(values = c("navy", "orange", "hotpink")) +
  geom_text(
    mapping = aes(label = y, !!!aes_contrast()),
    position = position_dodge2(width = 0.75, preserve = "single"),
    vjust = 1.33,
    show.legend = FALSE,
  )

}
