% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{guide_axis_logticks}
\alias{guide_axis_manual}
\alias{guide_axis_minor}
\alias{guide_axis_nested}
\alias{guide_axis_scalebar}
\alias{guide_axis_truncated}
\alias{guide_axis_colour}
\alias{guide_axis_color}
\alias{guide_dendro}
\alias{ggsubset}
\alias{scale_x_dendrogram}
\alias{scale_y_dendrogram}
\title{Deprecated functions}
\usage{
guide_axis_logticks(...)

guide_axis_manual(...)

guide_axis_minor(...)

guide_axis_nested(...)

guide_axis_scalebar(...)

guide_axis_truncated(...)

guide_axis_colour(...)

guide_axis_color(...)

guide_dendro(...)

ggsubset(...)

scale_x_dendrogram(...)

scale_y_dendrogram(...)
}
\arguments{
\item{...}{Not used.}
}
\value{
None, raises deprecation signal
}
\description{
The functions listed here are deprecated and no longer work.
}
\examples{
# None
}
