% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facetted_pos_scales.R
\name{facetted_pos_scales}
\alias{facetted_pos_scales}
\title{Set individual scales at facets}
\usage{
facetted_pos_scales(x = NULL, y = NULL)
}
\arguments{
\item{x}{A \code{list} wherein elements are either x position scales or
\code{NULL}s.}

\item{y}{A \code{list} wherein elements are either y position scales or
\code{NULL}s.}
}
\description{
Under default circumstances, using facets in a plot limits the freedom with
which the position scales (x and y) can be defined. Calling this function
after facets have been defined allows fine-tuning of the position scales.
}
\details{
It is intended that this function works with both
  \code{\link[ggplot2]{facet_wrap}} and \code{\link[ggplot2]{facet_grid}}.
  For \code{facet_wrap}, the scales are used for each individual panel. For
  \code{facet_grid}, the scales are used for the rows and columns. Note that
  these facets must be used with \code{scales = "free"} or \code{"free_x"} or
  \code{"free_y"}, depending on what scales are added.

  Axis titles are derived from the first scale in the list (or the default
  position scale when the first list element is \code{NULL}).

  It is allowed to use individual scale transformations for facets, but this
  functionality comes with the trade-off that the out of bounds (\code{oob})
  argument for individual scales is ignored. Values that are out of bounds
  will be clipped.

  \code{NULL}s are valid list elements and signal that the default position
  scale should be used at the position in the list where the \code{NULL}
  occurs. Since transformations are applied before facet scales are
  initiated, it is not recommended to use a default position scale with a
  transformation other than \code{trans = "identity"} (the default).
}
\examples{
# Reversing the y-axis in the second panel
ggplot(iris, aes(Sepal.Width, Sepal.Length)) +
  geom_point(aes(colour = Species)) +
  facet_wrap(Species ~ ., scales = "free_y") +
  facetted_pos_scales(
    y = list(NULL, scale_y_continuous(trans = "reverse"))
  )
}
\seealso{
\code{\link[ggplot2]{scale_continuous}} and \code{scale_x_discrete}.
}
