% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggh4x_extensions.R, R/coord_axes_inside.R,
%   R/facet_grid2.R, R/facet_wrap2.R, R/facet_manual.R, R/facet_nested.R,
%   R/facet_nested_wrap.R, R/geom_box.R, R/geom_pointpath.R,
%   R/geom_polygonraster.R, R/geom_rectrug.R, R/geom_text_aimed.R,
%   R/guide_stringlegend.R, R/position_disjoint_ranges.R,
%   R/position_lineartrans.R, R/scale_dendrogram.R, R/stat_difference.R,
%   R/stat_funxy.R, R/stat_rle.R, R/stat_roll.R, R/stat_theodensity.R,
%   R/strip_vanilla.R, R/strip_themed.R, R/strip_nested.R, R/strip_split.R
\docType{data}
\name{ggh4x_extensions}
\alias{ggh4x_extensions}
\alias{CoordAxesInside}
\alias{FacetGrid2}
\alias{FacetWrap2}
\alias{FacetManual}
\alias{FacetNested}
\alias{FacetNestedWrap}
\alias{GeomBox}
\alias{GeomPointPath}
\alias{GeomPointpath}
\alias{GeomPolygonRaster}
\alias{GeomRectMargin}
\alias{GeomTileMargin}
\alias{GeomTextAimed}
\alias{GuideStringlegend}
\alias{PositionDisjointRanges}
\alias{PositionLinearTrans}
\alias{ScaleDendrogram}
\alias{StatDifference}
\alias{StatFunxy}
\alias{StatRle}
\alias{StatRollingkernel}
\alias{StatTheoDensity}
\alias{Strip}
\alias{StripThemed}
\alias{StripNested}
\alias{StripSplit}
\title{ggh4x extensions to ggplot2}
\description{
ggh4x relies on the extension mechanism of ggplot2 through
ggproto class objects, which allows cross-package inheritance of objects
such as geoms, stats, facets, scales and coordinate systems. These objects
can be ignored by users for the purpose of making plots, since interacting
with these objects is preferred through various geom_, stat_, facet_,
coord_ and scale_ functions.
}
\seealso{
\link[ggplot2:ggproto]{ggproto}
}
\keyword{datasets}
