% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_dendrogram.R
\name{guide_dendro}
\alias{guide_dendro}
\title{Dendrogram guide}
\usage{
guide_dendro(
  title = waiver(),
  check.overlap = FALSE,
  n.dodge = 1,
  order = 0,
  position = waiver(),
  dendro = waiver()
)
}
\arguments{
\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link[ggplot2:waiver]{waiver()}}), the name of the scale object or the name
specified in \code{\link[ggplot2:labs]{labs()}} is used for the title.}

\item{check.overlap}{silently remove overlapping labels,
(recursively) prioritizing the first, last, and middle labels.}

\item{n.dodge}{The number of rows (for vertical axes) or columns (for
horizontal axes) that should be used to render the labels. This is
useful for displaying labels that would otherwise overlap.}

\item{order}{Used to determine the order of the guides (left-to-right,
top-to-bottom), if more than one  guide must be drawn at the same location.}

\item{position}{Where this guide should be drawn: one of top, bottom,
left, or right.}

\item{dendro}{Relevant plotting data for a dendrogram such as those returned
by \code{\link[ggdendro]{dendro_data}}.}
}
\description{
Visual representation of a discrete variable with hierarchical relationships
between members, like those detailed in
\code{\link[=scale_x_dendrogram]{scale_(x|y)_dendrogram)}}.
}
\examples{
clust <- hclust(dist(USArrests), "ave")

df <- cbind(State = rownames(USArrests),
            USArrests)
df <- reshape2::melt(df, id.vars = "State")

# The guide function can be used to customise the axis
ggplot(df, aes(variable, State, fill = value)) +
  geom_raster() +
scale_y_dendrogram(hclust = clust, 
                   guide = guide_dendro(n.dodge = 2))
}
