% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{rgb_pal}
\alias{rgb_pal}
\title{RGB Palette}
\usage{
rgb_pal(x)
}
\arguments{
\item{x}{An \code{rgb_spec} vector with values between 0-1.}
}
\value{
A \code{character} vector with colours.
}
\description{
Palette function for \code{scale_colour/fill_rgb()}.
}
\examples{
rgb_pal(rgb_spec(c(0, 1), c(0.5, 0.5), c(0.25, 0.75)))
}
