% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_colour_spec.R
\name{rgb_spec}
\alias{rgb_spec}
\title{RGB specification}
\usage{
rgb_spec(r = 0, g = 0, b = 0)
}
\arguments{
\item{r, g, b}{\code{numeric()} vectors with values for the red, green and blue
channels respectively.}
}
\value{
An \code{rgb_spec}, \code{vctrs_rcrd} S3 class.
}
\description{
This function creates a vector that stores values for red green and blue
channels in the RGB colourspace.
}
\examples{
# Creating an `rgb_spec` vector
rgb <- rgb_spec(r = c(1, 2, 3), g = c(40, 50, 60), b = c(-700, -800, -900))

# Rescaling vector to have values between 0-255
rgb <- scales::rescale(rgb, to = c(0, 255))

# Rescale vector to the 0-1 range is the default
rgb <- scales::rescale(rgb)

# Converting the rescaled vector to colours assumes 0-1 range
rgb_pal(rgb)

# Example usage as colour aesthetic
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = rgb_spec(mpg, cyl, drat)))
}
