% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_extensions}
\alias{theme_extensions}
\alias{ggh4x_theme_elements}
\title{Theme extensions}
\arguments{
\item{ggh4x.facet.nestline}{An \code{\link[=element_line]{element_line()}} used as the parent for the
\code{nest_line} argument in \code{\link[=facet_nested]{facet_nested()}} and \code{\link[=facet_nested_wrap]{facet_nested_wrap()}}.
Inherits directly from the '\code{line}' theme element.}

\item{ggh4x.axis.nestline, ggh4x.axis.nestline.x, ggh4x.axis.nestline.y}{An
\code{\link[=element_line]{element_line()}} used as the line to separate different layers of labels
in \code{\link[=guide_axis_nested]{guide_axis_nested()}}. Inherits from the '\code{axis.ticks}' theme element.}

\item{ggh4x.axis.nesttext.x, ggh4x.axis.nesttext.y}{An \code{\link[=element_text]{element_text()}} used
to differentiate text higher in the hierarchy from the axis labels directly
next to the axis line in \code{\link[=guide_axis_nested]{guide_axis_nested()}}. Inherits from the
'\code{axis.text.x}' and '\code{axis.text.y}' theme elements respectively.}

\item{ggh4x.axis.ticks.length.minor}{A \code{\link[=rel]{rel()}} object used to set the size
of minor tick marks relative to the regular tick marks. This is used in the
\code{\link[=guide_axis_minor]{guide_axis_minor()}} and \code{\link[=guide_axis_logticks]{guide_axis_logticks()}} functions. Defaults
to \code{rel(2/3)}.}

\item{ggh4x.axis.ticks.length.mini}{A \code{\link[=rel]{rel()}} object used to set the size
of the smallest tick marks relative to regular tick marks. This is only
used in the \code{\link[=guide_axis_logticks]{guide_axis_logticks()}} function. Defaults to \code{rel(1/3)}.}
}
\description{
Some functions in \pkg{ggh4x} are using extensions to the theme
system. These extended theme argument are listed below, along with what
elements they are expected to be, and in what function(s) they are used.
}
