% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggh4x-package.R
\docType{package}
\name{ggh4x-package}
\alias{ggh4x}
\alias{ggh4x-package}
\title{ggh4x: Hacks for ggplot2}
\description{
A ggplot2 extensions with some stats, geoms and hacks.
}
\author{
\strong{Maintainer}: Teun van den Brand \email{tahvdbrand@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
