% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggh4x-package.R
\docType{package}
\name{ggh4x-package}
\alias{ggh4x}
\alias{ggh4x-package}
\title{ggh4x: Hacks for 'ggplot2'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A 'ggplot2' extension that does a variety of little
    helpful things.  The package extends 'ggplot2' facets through
    customisation, by setting individual scales per panel, resizing panels
    and providing nested facets.  Also allows multiple colour and fill
    scales per plot. Also hosts a smaller collection of stats, geoms and axis 
    guides.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/teunbrand/ggh4x}
  \item Report bugs at \url{https://github.com/teunbrand/ggh4x/issues}
}

}
\author{
\strong{Maintainer}: Teun van den Brand \email{tahvdbrand@gmail.com} (\href{https://orcid.org/0000-0002-9335-7468}{ORCID})

}
\keyword{internal}
