% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_nested.R, R/facet_wrap2.R,
%   R/facet_nested_wrap.R, R/geom_pointpath.R, R/geom_polygonraster.R,
%   R/geom_rectrug.R, R/ggh4x_extensions.R, R/position_disjoint_ranges.R,
%   R/position_lineartrans.R, R/scale_dendrogram.R, R/stat_funxy.R,
%   R/stat_rle.R, R/stat_roll.R, R/stat_theodensity.R
\docType{data}
\name{FacetNested}
\alias{FacetNested}
\alias{FacetWrap2}
\alias{FacetNestedWrap}
\alias{GeomPointPath}
\alias{GeomPolygonRaster}
\alias{GeomRectMargin}
\alias{GeomTileMargin}
\alias{ggh4x_extensions}
\alias{PositionDisjointRanges}
\alias{PositionLinearTrans}
\alias{ScaleDendrogram}
\alias{StatFunxy}
\alias{StatRle}
\alias{StatRollingkernel}
\alias{StatTheoDensity}
\title{ggh4x extensions to ggplot2}
\description{
ggnomics relies on the extension mechanism of ggplot2 through
  ggproto class objects, which allows cross-package inheritance of objects
  such as geoms, stats, facets, scales and coordinate systems. These objects
  can be ignored by users for the purpose of making plots, since interacting
  with these objects is preferred through various geom_\*, stat_\*, facet_\*,
  coord_\* and scale_\* functions.
}
\seealso{
\link[ggplot2]{ggproto}
}
\keyword{datasets}
